commit for force dub

+++ build/controllers/userComposer.js
@@ -2,6 +2,7 @@ import { Composer } from "grammy";
 import { InlineKeyboard } from "grammy";
 import { insert_audio, insert_document, insert_user, insert_video, search_audio_file_id, search_document_file_id, search_video_file_id, } from "../functions/dbFunc.js";
 import { cleanFileName, extractSearchTerm, keyboardlist, userMode, } from "../functions/helperFunc.js";
+export const notAuthorized = new Set();
 export const userComposer = new Composer();
 userComposer.on("callback_query:data", async (ctx) => {
     var _a;
@@ -101,60 +102,63 @@ userComposer.on("callback_query:data", async (ctx) => {
 userComposer.chatType("private").command("start", async (ctx) => {
     try {
         if (ctx.match) {
-            const parts = ctx.match.split("_-_");
-            const file_unique_id = parts[1];
-            const type = parts[0];
-            const code = parts[2];
-            const currentHour = new Date().getHours().toString();
-            // if (code == encodedString) {
-            if (type == "doc") {
-                const { filteredDocs } = await search_document_file_id(file_unique_id);
-                if (filteredDocs.file_caption != "") {
-                    await ctx.replyWithDocument(filteredDocs.file_id, {
-                        caption: filteredDocs.file_caption,
-                        protect_content: true,
-                    });
+            const isMember = await ctx.api.getChatMember(Number(process.env.CHECKMEMBER), ctx.from.id);
+            if (["member"].includes(isMember.status)) {
+                const parts = ctx.match.split("_-_");
+                const file_unique_id = parts[1];
+                const type = parts[0];
+                if (type == "doc") {
+                    const { filteredDocs } = await search_document_file_id(file_unique_id);
+                    if (filteredDocs.file_caption != "") {
+                        await ctx.replyWithDocument(filteredDocs.file_id, {
+                            caption: filteredDocs.file_caption,
+                            protect_content: true,
+                        });
+                    }
+                    else {
+                        await ctx.replyWithDocument(filteredDocs.file_id, {
+                            caption: filteredDocs.file_name,
+                            protect_content: true,
+                        });
+                    }
                 }
-                else {
-                    await ctx.replyWithDocument(filteredDocs.file_id, {
-                        caption: filteredDocs.file_name,
-                        protect_content: true,
-                    });
+                else if (type == "vid") {
+                    const { filteredDocs } = await search_video_file_id(file_unique_id);
+                    if (filteredDocs.file_caption != "") {
+                        await ctx.replyWithDocument(filteredDocs.file_id, {
+                            caption: filteredDocs.file_caption,
+                            protect_content: true,
+                        });
+                    }
+                    else {
+                        await ctx.replyWithDocument(filteredDocs.file_id, {
+                            caption: filteredDocs.file_name,
+                            protect_content: true,
+                        });
+                    }
+                }
+                else if (type == "aud") {
+                    const { filteredDocs } = await search_audio_file_id(file_unique_id);
+                    if (filteredDocs.file_caption != "") {
+                        await ctx.replyWithDocument(filteredDocs.file_id, {
+                            caption: filteredDocs.file_caption,
+                            protect_content: true,
+                        });
+                    }
+                    else {
+                        await ctx.replyWithDocument(filteredDocs.file_id, {
+                            caption: filteredDocs.file_name,
+                            protect_content: true,
+                        });
+                    }
                 }
             }
-            else if (type == "vid") {
-                const { filteredDocs } = await search_video_file_id(file_unique_id);
-                if (filteredDocs.file_caption != "") {
-                    await ctx.replyWithDocument(filteredDocs.file_id, {
-                        caption: filteredDocs.file_caption,
-                        protect_content: true,
-                    });
-                }
-                else {
-                    await ctx.replyWithDocument(filteredDocs.file_id, {
-                        caption: filteredDocs.file_name,
-                        protect_content: true,
-                    });
+            else {
+                if (!notAuthorized.has(ctx.from.id)) {
+                    await ctx.reply("Looks like you're not authorized ❌. Please visit the channel @dedxec for more information.");
+                    notAuthorized.add(ctx.from.id);
                 }
             }
-            else if (type == "aud") {
-                const { filteredDocs } = await search_audio_file_id(file_unique_id);
-                if (filteredDocs.file_caption != "") {
-                    await ctx.replyWithDocument(filteredDocs.file_id, {
-                        caption: filteredDocs.file_caption,
-                        protect_content: true,
-                    });
-                }
-                else {
-                    await ctx.replyWithDocument(filteredDocs.file_id, {
-                        caption: filteredDocs.file_name,
-                        protect_content: true,
-                    });
-                }
-            }
-            // } else {
-            //   await ctx.reply("Your are not authorized ❌");
-            // }
         }
         else {
             if (ctx.from) {
@@ -167,7 +171,7 @@ userComposer.chatType("private").command("start", async (ctx) => {
                 await insert_user(data);
             }
             userMode.set(ctx.from.id, "document");
-            ctx.reply(`👋 Hi, I'm ${ctx.me.first_name}! 📄🎥🎵 I am a movie serving bot`, { parse_mode: "HTML" });
+            ctx.reply(`👋 Hi ${ctx.from.first_name}, I'm ${ctx.me.first_name}! 📄🎥🎵 Feel free to search for any media files through me 😊.`, { parse_mode: "HTML" });
         }
     }
     catch (error) { }
@@ -394,13 +398,48 @@ userComposer.chatType("private").on(":text", async (ctx, next) => {
     try {
         const msgDeleteTime = parseInt(process.env.MESSAGE_DELETE_TIME || "");
         if (!ctx.msg.text.includes("/")) {
-            if (userMode.get(ctx.from.id)) {
-                if (hasFiveParts(ctx.msg.text)) {
-                    // Create a task queue
-                    const taskQueue = new TaskQueue();
-                    // Define some tasks
-                    const task1 = () => new Promise(async (resolve) => {
-                        var _a, _b;
+            const isMember = await ctx.api.getChatMember(Number(process.env.CHECKMEMBER), ctx.from.id);
+            if (["member"].includes(isMember.status)) {
+                if (userMode.get(ctx.from.id)) {
+                    if (hasFiveParts(ctx.msg.text)) {
+                        // Create a task queue
+                        const taskQueue = new TaskQueue();
+                        // Define some tasks
+                        const task1 = () => new Promise(async (resolve) => {
+                            var _a, _b;
+                            setTimeout(async () => {
+                                try {
+                                    await ctx.deleteMessage();
+                                }
+                                catch (error) {
+                                    console.log(error);
+                                }
+                            }, msgDeleteTime);
+                            const inlineKeyboard = await keyboardlist(ctx, 1, ctx.msg.text);
+                            if (inlineKeyboard) {
+                                const { message_id } = await ctx.reply(`Hey <a href="tg://user?id=${(_a = ctx.from) === null || _a === void 0 ? void 0 : _a.id}">${(_b = ctx.from) === null || _b === void 0 ? void 0 : _b.first_name}</a> , You Searched For: <code>${ctx.msg.text}</code>`, {
+                                    reply_markup: inlineKeyboard,
+                                    parse_mode: "HTML",
+                                });
+                                setTimeout(async () => {
+                                    try {
+                                        await ctx.api.deleteMessage(ctx.chat.id, message_id);
+                                    }
+                                    catch (error) {
+                                        console.log(error);
+                                    }
+                                }, msgDeleteTime);
+                            }
+                            resolve();
+                        });
+                        // Enqueue tasks
+                        taskQueue.enqueue(task1);
+                        // Execute tasks in the queue
+                        taskQueue.execute().then(() => {
+                            console.log("task executed");
+                        });
+                    }
+                    else {
                         setTimeout(async () => {
                             try {
                                 await ctx.deleteMessage();
@@ -409,57 +448,31 @@ userComposer.chatType("private").on(":text", async (ctx, next) => {
                                 console.log(error);
                             }
                         }, msgDeleteTime);
-                        const inlineKeyboard = await keyboardlist(ctx, 1, ctx.msg.text);
-                        if (inlineKeyboard) {
-                            const { message_id } = await ctx.reply(`Hey <a href="tg://user?id=${(_a = ctx.from) === null || _a === void 0 ? void 0 : _a.id}">${(_b = ctx.from) === null || _b === void 0 ? void 0 : _b.first_name}</a> , You Searched For: <code>${ctx.msg.text}</code>`, {
-                                reply_markup: inlineKeyboard,
-                                parse_mode: "HTML",
-                            });
-                            setTimeout(async () => {
-                                try {
-                                    await ctx.api.deleteMessage(ctx.chat.id, message_id);
-                                }
-                                catch (error) {
-                                    console.log(error);
-                                }
-                            }, msgDeleteTime);
-                        }
-                        resolve();
-                    });
-                    // Enqueue tasks
-                    taskQueue.enqueue(task1);
-                    // Execute tasks in the queue
-                    taskQueue.execute().then(() => {
-                        console.log("task executed");
-                    });
+                        const { message_id } = await ctx.reply(`Please limit your request to 5 words or less.\n\neg: <code>Money Heist s04 1080p</code>`, {
+                            parse_mode: "HTML",
+                            reply_parameters: {
+                                message_id: ctx.msg.message_id,
+                            },
+                        });
+                        setTimeout(async () => {
+                            try {
+                                await ctx.api.deleteMessage(ctx.chat.id, message_id);
+                            }
+                            catch (error) {
+                                console.log(error);
+                            }
+                        }, msgDeleteTime);
+                    }
                 }
                 else {
-                    setTimeout(async () => {
-                        try {
-                            await ctx.deleteMessage();
-                        }
-                        catch (error) {
-                            console.log(error);
-                        }
-                    }, msgDeleteTime);
-                    const { message_id } = await ctx.reply(`Please limit your request to 5 words or less.\n\neg: <code>Money Heist s04 1080p</code>`, {
-                        parse_mode: "HTML",
-                        reply_parameters: {
-                            message_id: ctx.msg.message_id,
-                        },
-                    });
-                    setTimeout(async () => {
-                        try {
-                            await ctx.api.deleteMessage(ctx.chat.id, message_id);
-                        }
-                        catch (error) {
-                            console.log(error);
-                        }
-                    }, msgDeleteTime);
+                    await ctx.reply("Please set your search mode using /mode");
                 }
             }
             else {
-                await ctx.reply("Please set your search mode using /mode");
+                if (!notAuthorized.has(ctx.from.id)) {
+                    await ctx.reply("🤔 Looks like you're not authorized ❌. Please visit the channel @dedxec for more information.");
+                    notAuthorized.add(ctx.from.id);
+                }
             }
         }
     }
